mrc-cuda-graph-setup(extract-cuda-graph-setup-move,mechanical_provable): move init_cuda_graphs chain to cuda_graph_setup module (cut+paste)
PR-Title: Extract cuda-graph setup into a module
